ImageBundle and ClientBundle generator errors hinted about Class.getResource() whereas they don't support leading '/'.
The ImageBundle javadoc is clear about it, and there's even a test for it,
but the error message erroneously talked about Class.getResource() instead
of ClassLoader.getResource().

It's a different story for ClientBundle: we could easily make it accept
leading '/' in paths (and possibly skip the package-relative lookup in this
case; it would make it almost equivalent to Class.getResource() then, with
the exception of the "fallback to global lookup if package-relative failed"
behavior), I chosed to simply fix the message here. Note that in any case,
the message is really about the path being *formatted* as one or the other
getResource() would expect, not about the lookup algorithm (which BTW isn't
clearly documented anywhere).

Fixes issue 7375

Change-Id: I4b3bbb1173841d40216981c93515026c156cfa67
Review-Link: https://gwt-review.googlesource.com/#/c/1790/

Review by: goktug@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@11478 8db76d5a-ed1c-0410-87a9-c151d255dfc7